+++ NEWS.md
@@ -10,6 +10,7 @@ xxxx-xx-xx
   - Fix IsSimpleOp for MultiPoint with empty element (GH-1005, Martin Davis)
   - Fix PreparedPolygonContains for GC with MultiPoint (GH-1008, Martin Davis)
   - Fix TopologyPreservingSimplifier to prevent jumping components (GH-1012, Martin Davis)
+  - Fix reading WKT with EMPTY token with white space (GH-1025, Mike Taves)
 
 ## Changes in 3.12.1
 2023-11-11